Funciones terminadas y documentadas
